test: move multi-identity cert coverage to its own file
